Scaling of OBC during sampling
The OBC force field can be scaled. The OBC_ligand option was removed
and replaced with the —solvation option, which can be “Full”,
“Desolvated”, or “Fractional”. With “Full” solvation, the OBC force
field is always on. With “Desolvated”, it is on for the ligand alone at
T=T_TARGET and off with T=T_HIGH and off for docking. “Fractional”
solvation is still being implemented.