Fixes to not reject end points, better estimate thermodynamic length, and to strengthen the pose restraint
The feature of rejecting previous states during initialization was
removed for the initial state.

Thermodynamic length for desolvated had an erroneous OBC term. Also for
cooling it was not multiplied by deltaT as it should have been,
particularly with scaled solvation.

The default pose restraint was strengthened from 200 kJ/mol/K to 2000
kJ/mol/K.